refactor: deduplicate AllCredentialsFailedError and fix logging in switcher
