+++ openviking/models/embedder/base.py
@@ -15,6 +15,7 @@ from openviking.utils.embedding_input import (
     resolve_embedding_max_input_tokens,
     truncate_embedding_input,
 )
+from openviking.utils.exceptions import AllCredentialsFailedError
 from openviking.utils.model_retry import (
     OrderedCredentialSwitcher,
     classify_api_error,
@@ -719,22 +720,6 @@ def exponential_backoff_retry(
                 )
 
             time.sleep(delay)
-
-
-class AllCredentialsFailedError(Exception):
-    """Raised when all credentials in the chain have failed."""
-
-    def __init__(self, errors: list[tuple[str, str, Exception, int]]):
-        """Initialize the error with a list of credential failures.
-
-        Args:
-            errors: List of tuples containing (credential_id, error_class, exception, attempts)
-        """
-        self.errors = errors
-        message = "All credentials failed:\n" + "\n".join(
-            f"  - {cred_id}: {error_class} - {exc}" for cred_id, error_class, exc, attempts in errors
-        )
-        super().__init__(message)
 
 
 class FailoverEmbedder(EmbedderBase):
