+++ openviking/utils/model_retry.py
@@ -8,6 +8,8 @@ import threading
 import time
 from typing import Awaitable, Callable, TypeVar
 
+from openviking.utils.exceptions import AllCredentialsFailedError
+
 logger = logging.getLogger(__name__)
 
 T = TypeVar("T")
@@ -445,7 +447,7 @@ class OrderedCredentialSwitcher:
                 self._active_idx = min(self._active_idx + 1, self._n)
                 self._last_switch_time = time.monotonic()
                 self._active_request_count = 0
-            logger.warning(f"Credential {idx} failed with unknown error class: {error_class}, advancing to {self._active_idx}")
+                logger.warning(f"Credential {idx} failed with unknown error class: {error_class}, advancing to {self._active_idx}")
             return True
 
     @property
